app/api/requirements_api.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# Async support
aiofiles>=23.0.0

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# MongoDB
motor>=3.3.0  # Async MongoDB driver
pymongo>=4.6.0  # Sync MongoDB driver
dnspython>=2.4.0  # For MongoDB SRV records

# Environment and Configuration
python-dotenv>=1.0.0
